9/19/2023

+++ Assets/Scenes/Lan/Enemy/Lan Mobs Melee.cs
@@ -298,7 +298,7 @@ public class LanMobsMelee : NetworkBehaviour
         Transform temp;
         temp = damagePool.GetChild(0);
         temp.GetComponent<TextMeshProUGUI>().color = Color.red;
-        temp.GetComponent<TextMeshProUGUI>().SetText(damage.ToString());
+        temp.GetComponent<TextMeshProUGUI>().SetText(((int)damage).ToString());
         temp.SetParent(transform);
         temp.gameObject.SetActive(true);
